refactor(gui-mcp): decouple JSON and DX paths into separate modules
Reorganize src/ into json/, dx/, and shared/ so the two authoring paths
read as independent units. json/ and dx/ no longer import each other; the
only shared code they cross into is shared/ (the reactive-expression
linter both reuse, plus the tool-registry helper)
