+++ libs/gui/mcp/src/json/generate-from-json-schema.spec.ts
