+++ libs/gui/mcp/src/json/generate-from-openapi.spec.ts
