+++ libs/gui/mcp/src/json/get-widget-spec.spec.ts
