+++ libs/gui/mcp/src/json/mapping/json-schema-to-gui.ts
