+++ libs/gui/mcp/src/json/mapping/validator.ts
