+++ libs/gui/mcp/src/json/schemas/index.ts
