+++ libs/gui/mcp/src/json/schemas/shallow-validators.ts
