+++ libs/gui/mcp/src/json/string-interpolation.spec.ts
