+++ libs/gui/mcp/src/json/string-interpolation.ts
