+++ libs/gui/mcp/src/json/validate-form-definition.spec.ts
