+++ libs/gui/mcp/src/shared/get-concept.ts
