+++ libs/gui/mcp/src/shared/lint/reactive-expressions.ts
